Extract slack package
