+++ access/slack/slack_test.go
@@ -1,4 +1,4 @@
-package main
+package slack
 
 import (
 	"context"
